+++ e2e/__tests__/failures.test.ts
@@ -6,6 +6,7 @@
  */
 
 import * as path from 'path';
+import {isJestJasmineRun} from '@jest/test-utils';
 import {extractSummary, runYarnInstall} from '../Utils';
 import runJest from '../runJest';
 
@@ -64,7 +65,10 @@ test('works with async failures', () => {
   // Remove replacements when jasmine is gone
   const result = normalizeDots(rest)
     .replace(/.*thrown:.*\n/, '')
-    .replace(/.*Use jest\.setTimeout\(newTimeout\).*/, '<REPLACED>')
+    .replace(
+      /.*Add a timeout value to this test to increase the timeout, if this is a long-running test. See https:\/\/jestjs.io\/docs\/api#testname-fn-timeout..*/,
+      '<REPLACED>',
+    )
     .replace(/.*Timeout - Async callback was not.*/, '<REPLACED>');
 
   expect(result).toMatchSnapshot();
@@ -89,6 +93,23 @@ test('works with snapshot failures with hint', () => {
     result.substring(0, result.indexOf('Snapshot Summary')),
   ).toMatchSnapshot();
 });
+
+(isJestJasmineRun() ? test.skip : test)('works with error with cause', () => {
+  const {stderr} = runJest(dir, ['errorWithCause.test.js']);
+  const summary = normalizeDots(cleanStderr(stderr));
+
+  expect(summary).toMatchSnapshot();
+});
+
+(isJestJasmineRun() ? test.skip : test)(
+  'works with error with cause thrown outside tests',
+  () => {
+    const {stderr} = runJest(dir, ['errorWithCauseInDescribe.test.js']);
+    const summary = normalizeDots(cleanStderr(stderr));
+
+    expect(summary).toMatchSnapshot();
+  },
+);
 
 test('errors after test has completed', () => {
   const {stderr} = runJest(dir, ['errorAfterTestComplete.test.js']);